chore: move dataset fragment to dataset upload module
