+++ src/app/modules/dataset-module/component/datasetSubHeaderToolbar.tsx
@@ -21,7 +21,7 @@ import { useStoreActions, useStoreState } from "app/state/store/hooks";
 /** Project */
 import { styles } from "app/modules/dataset-module/component/styles";
 import DeleteDatasetDialog from "app/components/Dialogs/deleteDatasetDialog";
-import { ISnackbarState } from "app/fragments/datasets-fragment/upload-steps/previewFragment";
+import { ISnackbarState } from "app/modules/dataset-upload-module/upload-steps/previewFragment";
 import { InfoSnackbar } from "app/modules/report-module/components/reportSubHeaderToolbar";
 import { homeDisplayAtom } from "app/state/recoil/atoms";
 import { useRecoilState } from "recoil";
